+++ 3172/3/.gitignore
@@ -1,1 +1,3 @@
-db/flamp.db
+/db/*.sqlite
+/db/*.sqlite3-journal
+
