+++ 3172/3/Gemfile
@@ -7,5 +7,6 @@ gem 'sinatra-contrib'
 gem 'sqlite3'
 
 group :development do
+  gem 'pry-byebug'
   gem 'tux'
 end
