+++ 3172/3/Rakefile
@@ -1,2 +1,4 @@
-require "./app"
-require "sinatra/activerecord/rake"
+require_relative 'config/environment'
+require 'sinatra/activerecord/rake'
+
+# Type `rake -T` on your command line to see the available rake tasks.
