+++ 3172/3/app/controllers/application_controller.rb
@@ -1,0 +1,19 @@
+require 'sinatra/flash'
+require 'sinatra/namespace'
+require './config/environment'
+require_relative 'users_controller'
+
+class ApplicationController < Sinatra::Base
+  configure do
+    enable :sessions
+    set :views, 'app/views'
+  end
+
+  register Sinatra::ActiveRecordExtension
+  register Sinatra::Flash
+
+  get '/' do
+    @places = Place.all
+    erb :index
+  end
+end
