+++ 3172/3/app/models/place.rb
@@ -1,0 +1,3 @@
+class Place < ActiveRecord::Base
+  has_many :reviews
+end
