+++ 3172/3/app/models/review.rb
@@ -1,0 +1,3 @@
+class Review < ActiveRecord::Base
+  belongs_to :place
+end
