+++ 3172/3/app/models/user.rb
@@ -1,0 +1,2 @@
+class User < ActiveRecord::Base
+end
