+++ 3172/3/config/database.yml
@@ -1,0 +1,17 @@
+development:
+  adapter: sqlite3
+  database: db/flamp.db
+  pool: 5
+  timeout: 5000
+
+test:
+  adapter: sqlite3
+  database: db/flamp.db
+  pool: 5
+  timeout: 5000
+
+production:
+  adapter: sqlite3
+  database: db/flamp.db
+  pool: 5
+  timeout: 5000
