+++ 3172/3/db/migrate/20190726100603_create_table_places.rb
@@ -1,3 +1,5 @@
+# :reek:FeatureEnvy
+# :reek:UncommunicativeVariableName
 class CreateTablePlaces < ActiveRecord::Migration[5.2]
   def change
     create_table :places do |t|
