+++ 3172/3/db/migrate/20190728171721_create_table_users.rb
@@ -1,3 +1,5 @@
+# :reek:FeatureEnvy
+# :reek:UncommunicativeVariableName
 class CreateTableUsers < ActiveRecord::Migration[5.2]
   def change
     create_table :users do |t|
