+++ 3172/3/db/migrate/20190728203754_create_table_review.rb
@@ -1,3 +1,6 @@
+# :reek:FeatureEnvy
+# :reek:TooManyStatements
+# :reek:UncommunicativeVariableName
 class CreateTableReview < ActiveRecord::Migration[5.2]
   def change
     create_table :reviews do |t|
